+++ lib/hexagon_widget.dart
@@ -1,45 +1,108 @@
 library hexagon;
 
 import 'package:flutter/material.dart';
-import 'package:hexagon/hexagon_shadows_painter.dart';
-import 'package:hexagon/hexagon_type.dart';
+import 'package:flutter/rendering.dart';
 
 import 'hexagon_clipper.dart';
+import 'hexagon_painter.dart';
+import 'hexagon_type.dart';
 
 class HexagonWidget extends StatelessWidget {
   const HexagonWidget({
     Key key,
     this.width,
     this.height,
-    this.elevation = 2,
+    this.color,
     this.child,
-    this.type = HexagonType.FLAT,
-  })  : assert((elevation ?? 0) >= 0),
+    this.elevation = 2,
+    this.inBounds = true,
+    @required this.type,
+  })  : assert(width != null || height != null),
+        assert((elevation ?? 0) >= 0),
+        assert(type != null),
         super(key: key);
+
+  HexagonWidget.flat(
+      {this.width,
+      this.height,
+      this.color,
+      this.child,
+      this.elevation = 2,
+      this.inBounds = true})
+      : assert(width != null || height != null),
+        assert((elevation ?? 0) >= 0),
+        this.type = HexagonType.FLAT;
+
+  HexagonWidget.pointy(
+      {this.width,
+      this.height,
+      this.color,
+      this.child,
+      this.elevation = 2,
+      this.inBounds = true})
+      : assert(width != null || height != null),
+        assert((elevation ?? 0) >= 0),
+        this.type = HexagonType.POINTY;
 
   final HexagonType type;
   final double width;
   final double height;
   final double elevation;
+  final bool inBounds;
   final Widget child;
+  final Color color;
 
-  double get _ratio {
-    if (type == HexagonType.FLAT) return 1.1547005 / 1;
-    return 1 / 1.1547005;
+  Size _innerSize() {
+    var flatFactor = type.flatFactor(inBounds);
+    var pointyFactor = type.pointyFactor(inBounds);
+
+    if (height != null && width != null) return Size(width, height);
+    if (height != null)
+      return Size((height / type.ratio) * flatFactor / pointyFactor, height);
+    if (width != null)
+      return Size(width, (width * type.ratio) / flatFactor * pointyFactor);
+    return null;
+  }
+
+  Size _contentSize() {
+    var flatFactor = type.flatFactor(inBounds);
+    var pointyFactor = type.pointyFactor(inBounds);
+
+    if (height != null && width != null) return Size(width, height);
+    if (height != null)
+      return Size((height / type.ratio) / pointyFactor, height / pointyFactor);
+    if (width != null)
+      return Size(width / flatFactor, (width * type.ratio) / flatFactor);
+    return null;
   }
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      height: height,
-      width: width,
-      child: CustomPaint(
-        painter: HexagonShadowsPainter(type, elevation: elevation),
-        child: AspectRatio(
-          aspectRatio: _ratio,
+    var innerSize = _innerSize();
+    var contentSize = _contentSize();
+
+    return Align(
+      child: Container(
+        width: innerSize.width,
+        height: innerSize.height,
+        child: CustomPaint(
+          painter: HexagonPainter(
+            type,
+            color: color,
+            elevation: elevation,
+            inBounds: inBounds,
+          ),
           child: ClipPath(
-            clipper: HexagonClipper(type),
-            child: Container(color: Colors.white, child: child),
+            clipper: HexagonClipper(type, inBounds: inBounds),
+            child: OverflowBox(
+              alignment: Alignment.center,
+              maxHeight: contentSize.height,
+              maxWidth: contentSize.width,
+              child: Align(
+                alignment: Alignment.center,
+                child: child,
+              ),
+            ),
           ),
         ),
       ),