+++ lib/utils.dart
@@ -19,14 +19,17 @@ class HexagonUtils {
         center.dx + size * cos(angleRad), center.dy + size * sin(angleRad));
   }
 
+  /// Calculates hexagon corners for given size and with given center.
   static List<Point> flatHexagonCornerList(Offset center, double size) {
     List<Point> corners = List(6);
     for (int i = 0; i < 6; i++) {
       corners[i] = flatHexagonCorner(center, size, i);
     }
+
     return corners;
   }
 
+  /// Calculates hexagon corners for given size and with given center.
   static List<Point> pointyHexagonCornerList(Offset center, double size) {
     List<Point> corners = List(6);
     for (int i = 0; i < 6; i++) {
@@ -35,14 +38,18 @@ class HexagonUtils {
     return corners;
   }
 
-  static Path hexagonPath(Size size, HexagonType type) {
+  /// Returns path in shape of hexagon.
+  static Path hexagonPath(Size size, HexagonType type, {bool inBounds}) {
+    inBounds = inBounds == true;
     final center = Offset(size.width / 2, size.height / 2);
+
     List<Point> cornerList;
     if (type == HexagonType.FLAT) {
-      cornerList = HexagonUtils.flatHexagonCornerList(center, size.width / 2);
+      cornerList = HexagonUtils.flatHexagonCornerList(
+          center, size.width / type.flatFactor(inBounds) / 2);
     } else {
-      cornerList =
-          HexagonUtils.pointyHexagonCornerList(center, size.height / 2);
+      cornerList = HexagonUtils.pointyHexagonCornerList(
+          center, size.height / type.pointyFactor(inBounds) / 2);
     }
 
     final path = Path();
@@ -53,7 +60,6 @@ class HexagonUtils {
         path.lineTo(point.x, point.y);
       }
     });
-    path.close();
-    return path;
+    return path..close();
   }
 }